jQuery find() requires '#' to find elements by id
Much of this code was not working because jquery find() needed a '#' before the id. Multiple instances were being created when "+1" was passed to the badger.  This is also why the badger-text and badger-number classes were never being applied.  The leading '#' in the self generated id was actually being put in the elements id.  The css should be adjusted also because the badger- number and badger-text classes are now being applied.